+++ elwikijsp.root/bundles_elwiki/org.elwiki.data.persist/src/org/elwiki/data/persist/json/converters/PageContentConverter.java
@@ -48,7 +48,7 @@ public class PageContentConverter extends DeserialiseStuff
 
 		PageContent pageContent = Elwiki_dataFactory.eINSTANCE.createPageContent();
 
-		short version = getShort(jsonObject, VERSION);
+		int version = getInt(jsonObject, VERSION);
 		Date creationDate = getDate(jsonObject, CREATION_DATE, context);
 		String author = getString(jsonObject, AUTHOR);
 		String changeNote = getString(jsonObject, CHANGE_NOTE);
