+++ elwikijsp.root/bundles_elwiki_model/org.elwiki.data/gen-src/org/elwiki_data/impl/PageContentImpl.java
@@ -70,8 +70,8 @@ public class PageContentImpl extends CDOObjectImpl implements PageContent {
 	 * @generated
 	 */
 	@Override
-	public short getVersion() {
-		return (Short)eGet(Elwiki_dataPackage.Literals.IHISTORY_INFO__VERSION, true);
+	public int getVersion() {
+		return (Integer)eGet(Elwiki_dataPackage.Literals.IHISTORY_INFO__VERSION, true);
 	}
 
 	/**
@@ -80,7 +80,7 @@ public class PageContentImpl extends CDOObjectImpl implements PageContent {
 	 * @generated
 	 */
 	@Override
-	public void setVersion(short newVersion) {
+	public void setVersion(int newVersion) {
 		eSet(Elwiki_dataPackage.Literals.IHISTORY_INFO__VERSION, newVersion);
 	}
